ref(project): rename some files and fodlers to 'contributions'
