+++ src/client/src/app/app.module.ts
@@ -13,7 +13,7 @@ import { NavigationComponent } from './components/navigation/navigation.componen
 import { HeaderComponent } from './components/contents/header/header/header.component';
 import { AboutComponent } from './components/contents/about/about/about.component';
 import { SkillsComponent } from './components/contents/skills/skills/skills.component';
-import { ProjectsComponent } from './components/contents/projects/projects/projects.component';
+import { ProjectsComponent } from './components/contents/contributions/contributions/projects.component';
 import { ContactComponent } from './components/contents/contact/contact/contact.component';
 import { ContentsComponent } from './pages/contents/contents.component';
 import { ModalComponent } from './components/modal/modal.component';
