+++ src/client/src/app/components/contents/contributions/contributions/projects.component.html
@@ -1,6 +1,6 @@
 <div id="projects-section-page">
   <div class="projects-title-section">
-    <h3 class="projects-title">Projects</h3>
+    <h3 class="projects-title">Contributions</h3>
   </div>
   <div class="project-blocks-section">
