+++ src/client/src/app/components/contents/contributions/contributions/projects.component.scss
